% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{circle_area}
\alias{circle_area}
\title{This function calculates the area of a circle given radius}
\usage{
circle_area(r, digit)
}
\arguments{
\item{r}{Radius of the round}

\item{digit}{specify the accuracy of the output to which digit after decimal mark}
}
\description{
This function calculates the area of a circle given radius
}
\examples{
\dontrun{
circle_area(r=4, digit=2)
}
}
\keyword{helloworld}

