% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{dados}
\alias{dados}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame que contien mtchas colunas
\describe{
  \item{mpg}{moderno preco gole}
  \item{hp}{Harry potter}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
dados
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

