% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello}
\usage{
hello(i)
}
\arguments{
\item{i}{i é o parâmetro que a função pra usar (no caso nem tem)}
}
\description{
Printa a palavra hello
}
\details{
Destalhes ftw
}
\examples{
hello()

}

