\name{helloWorld-package}
\alias{helloWorld-package}
\docType{package}
\title{A sleek and modern hello world function.
}
\description{
helloWorld contains one function, \code{helloWorld} that is a sleek and modern hello world function.
}

\details{
\tabular{ll}{
Package: \tab helloWorld\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-4-2\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}

\author{Damian W. Betebenner \email{DBetebenner@nciea.org}} 

\keyword{package}
