% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{helloworld}
\alias{helloworld}
\title{Printing "hello world"}
\usage{
helloworld()
}
\value{
The function returns: "hello world"
}
\description{
This function just prints "hello world", nothing else
}
\examples{
helloworld()
}
