% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lmcum.r
\name{lmcum}
\alias{lmcum}
\title{A simple function for testing}
\arguments{
\item{formula}{A input formula}

\item{data}{A input data}
}
\value{
A linear model
}
\description{
A simple function for testing
}
\examples{
data(mtcars)
mtcars$cyl <- factor(mtcars$cyl)
mod <- lmcum(mpg ~ ., data = mtcars)
summary(mod)
}

