% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunction1.r
\name{myFunction1}
\alias{myFunction1}
\title{function to display a message}
\usage{
myFunction1(x = NULL)
}
\arguments{
\item{x}{a parameter to test whether the function receives a value}
}
\description{
function to display a message
}
\examples{
myFunction1("test parameter")

}
