\name{helloWorld-package}
\alias{helloWorld-package}
\alias{helloWorld}
\docType{package}
\title{
What the package does (short line)
  Send out greetings.
}
\description{
  Say hello to someone.
}
\details{
\tabular{ll}{
Package: \tab helloWorld\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-24\cr
License: \tab GPL\cr
}
}
\author{
Matt Furia <matt.furia@sagebase.org>

Maintainer: Matt Furia <matt.furia@sagebase.org>
}
\keyword{ package }
\examples{
  hello()
  hello("Matt")
}
