% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloworld.R
\name{helloworld}
\alias{helloworld}
\title{Hello World}
\usage{
helloworld(x)
}
\arguments{
\item{x}{string argument}
}
\value{
String output containing "Hello World" and the argument used for x
}
\description{
Creates a string output containing "Hello World" plus a string argument
}
\examples{
 
helloworld("wazup")

}
