% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloworld.R
\name{helloworld}
\alias{helloworld}
\title{It says Hello world to you.}
\usage{
helloworld()
}
\value{
invisibily return "Hello world".
}
\description{
It prints Hello world and return it invisibly.
}
\examples{
library(helloworld)
helloworld()
# Hello world
}
