\name{helloWorld}
\alias{helloWorld}
\title{hello world function}
\description{
A sleek and modern hello world function.
}
\usage{
helloWorld(n                                                    
            )                 
}

\arguments{
  \item{n}{An integer indicating how many times to repeat "Hello World". 
  }
} 


\details{Typical use of the function is supply an integer n to get "Hello World" repeated n times.
}

\value{Function returns "Hello World" n times. 
}

\author{Damian W. Betebenner \email{dbetebenner@nciea.org} }

\seealso{\code{\link{helloWorldData}}}

\examples{
helloWorld(10)  ## Returns "Hello World" 10 times
}

\keyword{ misc }
\keyword{ models }
