% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.holdout.cv.r
\name{one.holdout.cv}
\alias{one.holdout.cv}
\title{one.holdout.cv}
\usage{
one.holdout.cv(dataset, percentTrain = 0.7)
}
\arguments{
\item{dataset}{Matriz de datos con n muestras (filas)
y m variables (columnas)}

\item{percentTrain}{Numero decimal que corresponde al porcentaje
de muestras deseadas para entrenamiento.}
}
\value{
Una lista con dos matrices. training, la matriz de entrenamiento
   y testing, la matriz de prueba.
}
\description{
Funcion que recibe un dataset y retorna los sub datasets
correspondientes a entrenamiento (training) y pruebas (testing)
}
\examples{

library(tidyverse)
data <- one.holdout.cv(dataset = iris
                       , percentTrain = 0.75)
View(data$training)
View(data$testing)
}
\author{
Andres Mauricio Cifuentes Bernal
}
